Create LoginFINAL
My final and finished Login code started with a mix of mine and Eric's original code but ended up changing it and using chatgpt for some guidance. This is the final product. It prompts the user to login or register. It should save their login/registration data and should create a .txt file.